test_1/LD-WebTests/requirements.txt
alabaster==0.7.12
allure-pytest==2.8.40
allure-python-commons==2.8.40
apipkg==3.0.1
attrs==20.3.0
Babel==2.9.0
cairocffi==1.2.0
CairoSVG==2.5.2
certifi>=2021.10.8
cffi==1.14.5
chardet==4.0.0
cssselect2==0.4.1
defusedxml==0.7.1
Deprecated==1.2.10
docutils==0.16
execnet==1.8.0
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
Jinja2==2.11.3
MarkupSafe==1.1.1
multi-key-dict==2.0.3
packaging==20.9
pandas==1.3.5
pbr==5.5.1
Pillow==8.2.0
pluggy==0.13.1
py==1.11.0
pycparser==2.20
pydantic==1.8.2
Pygments==2.8.1
pyparsing==2.4.7
pytest<7.0.0, >=6.0.0
pytest-base-url<3.0.0, >=2.0.0
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-repeat==0.9.1
pytest-rerunfailures==9.1.1
pytest-selenium==4.0.0
pytest-timeout==2.1.0
pytest-variables<3.0.0, >=2.0.0
pytest-xdist==2.2.1
python-jenkins==1.7.0
pytz==2021.1
PyYAML==5.4.1
requests<3.0.0, >=2.26.0
selenium==4.8.3
six==1.15.0
snowballstemmer==2.1.0
Sphinx==3.5.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tenacity==6.3.1
tinycss2==1.1.0
toml==0.10.2
urllib3==1.26.4
webencodings==0.5.1
wrapt==1.12.1
yapf==0.31.0
xlrd==1.2.0
openpyxl==3.0.9